refactor: move process properties, methods to ProcessInfo refactor: add method PS_PROTECTION.ToString() refactor: add static property NativeMethods.Processes
