+++ deadlock-dotnet-sdk/Windows.Win32/PS_PROTECTION.cs
@@ -38,4 +38,29 @@ public struct PS_PROTECTION : IEquatable<PS_PROTECTION>
     public static bool operator ==(PS_PROTECTION left, PS_PROTECTION right) => left.Equals(right);
 
     public static bool operator !=(PS_PROTECTION left, PS_PROTECTION right) => !(left == right);
+
+    /// <summary>
+    /// Produces output similar to "Light (Antimalware)"
+    /// </summary>
+    /// <returns></returns>
+    public override string ToString()
+    {
+        string sType = string.Empty;
+        string sSigner = Signer is PS_PROTECTED_SIGNER.PsProtectedSignerNone ? string.Empty : ('(' + Signer.ToString().Replace("PsProtectedSigner", null) + ')');
+
+        switch (Type)
+        {
+            case PS_PROTECTED_TYPE.PsProtectedTypeNone:
+                sType = "None";
+                break;
+            case PS_PROTECTED_TYPE.PsProtectedTypeProtectedLight:
+                sType = "Light";
+                break;
+            case PS_PROTECTED_TYPE.PsProtectedTypeProtected:
+                sType = "Full";
+                break;
+        }
+
+        return $"{sType} {sSigner}";
+    }
 }
